created data communication (repository & viewmodel)

+++ app/src/main/java/com/babakan/cashier/data/repository/user/UserRepository.kt
@@ -1,8 +1,9 @@
-package com.babakan.cashier.data.repository.auth
+package com.babakan.cashier.data.repository.user
 
 import com.babakan.cashier.presentation.authentication.model.UserModel
 import com.babakan.cashier.utils.constant.RemoteData
 import com.babakan.cashier.data.state.UiState
+import com.babakan.cashier.presentation.owner.model.CategoryModel
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ListenerRegistration
 import com.google.firebase.firestore.SetOptions
@@ -16,7 +17,23 @@ class UserRepository(
 ) {
     private val userCollection = firestore.collection(RemoteData.COLLECTION_USERS)
 
-    suspend fun getUserDocument(
+    suspend fun getUsers(): UiState<List<UserModel>> {
+        return try {
+            val snapshot = userCollection
+                .get()
+                .await()
+
+            val users = snapshot.documents.mapNotNull { doc ->
+                if (doc.exists()) UserModel.fromDocumentSnapshot(doc) else null
+            }
+
+            UiState.Success(users)
+        } catch (e: Exception) {
+            UiState.Error("Terjadi kesalahan", e.message.toString())
+        }
+    }
+
+    suspend fun getUserById(
         userId: String
     ): UiState<UserModel> {
         return try {
@@ -26,8 +43,8 @@ class UserRepository(
                 .await()
 
             if (snapshot.exists()) {
-                val userModel = UserModel.fromDocumentSnapshot(snapshot)
-                UiState.Success(userModel)
+                val user = UserModel.fromDocumentSnapshot(snapshot)
+                UiState.Success(user)
             } else {
                 UiState.Error("Pengguna tidak ditemukan", "Pengguna dengan ID $userId tidak ditemukan")
             }
@@ -36,26 +53,27 @@ class UserRepository(
         }
     }
 
-    suspend fun setUserDocument(
+    suspend fun setUserById(
         userId: String,
         userModel: UserModel
-    ): UiState<String> {
+    ): UiState<Unit> {
         return try {
             userCollection
                 .document(userId)
                 .set(userModel.toJson(), SetOptions.merge())
                 .await()
-            UiState.Success("Data pengguna berhasil disimpan")
+
+            UiState.Success(Unit)
         } catch (e: Exception) {
             UiState.Error("Terjadi kesalahan", e.message.toString())
         }
     }
 
-    suspend fun updateUserDocument(
+    suspend fun updateUserById(
         userId: String,
         fieldName: String,
         newValue: Any
-    ): UiState<String> {
+    ): UiState<Unit> {
         return try {
             val updatedField = mapOf(fieldName to newValue)
 
@@ -64,27 +82,54 @@ class UserRepository(
                 .update(updatedField)
                 .await()
 
-            UiState.Success("'$fieldName' berhasil diperbarui")
+            UiState.Success(Unit)
         } catch (e: Exception) {
             UiState.Error("Terjadi kesalahan", e.message.toString())
         }
     }
 
-    suspend fun deleteUserDocument(
+    suspend fun deleteUserById(
         userId: String
-    ): UiState<String> {
+    ): UiState<Unit> {
         return try {
-            userCollection
-                .document(userId)
-                .delete()
-                .await()
-            UiState.Success("Data pengguna berhasil dihapus")
+            val updateState = updateUserById(userId, RemoteData.FIELD_IS_ACTIVE, false)
+
+            if (updateState is UiState.Success) {
+                userCollection
+                    .document(userId)
+                    .delete()
+                    .await()
+
+                UiState.Success(Unit)
+            } else {
+                UiState.Error("Gagal menghapus pengguna", "Gagal menonaktifkan pengguna sebelum menghapus")
+            }
         } catch (e: Exception) {
             UiState.Error("Terjadi kesalahan", e.message.toString())
         }
     }
 
-    fun observeIsActiveField(
+    suspend fun searchUsersByName(
+        query: String
+    ): UiState<List<UserModel>> {
+        return try {
+            val snapshot = userCollection
+                .whereGreaterThanOrEqualTo(RemoteData.FIELD_NAME, query)
+                .whereLessThanOrEqualTo(RemoteData.FIELD_NAME, query + "\uf8ff")
+                .get()
+                .await()
+
+            val users = snapshot.documents.mapNotNull { doc ->
+                if (doc.exists()) UserModel.fromDocumentSnapshot(doc) else null
+            }
+
+            UiState.Success(users)
+        } catch (e: Exception) {
+            UiState.Error("Terjadi kesalahan", e.message.toString())
+        }
+    }
+
+    fun listenIsActiveById(
         userId: String
     ): Flow<Boolean> = callbackFlow {
         val listenerRegistration: ListenerRegistration = userCollection
@@ -94,7 +139,6 @@ class UserRepository(
                     close(error)
                     return@addSnapshotListener
                 }
-
                 if (snapshot != null && snapshot.exists()) {
                     val isActive = snapshot.getBoolean(RemoteData.FIELD_IS_ACTIVE) ?: false
                     trySend(isActive)
